Atualiza master com develop

+++ app/Http/Controllers/WorkspaceController.php
@@ -11,6 +11,7 @@ use App\Http\Resources\WorkspaceResource;
 use App\Models\Goal;
 use App\Models\BoardList;
 use App\Constants\BoardListsKeys;
+use App\Constants\CardTypes;
 
 class WorkspaceController extends Controller
 {
@@ -80,10 +81,19 @@ class WorkspaceController extends Controller
 
 	private function createBoardList($key, $workspace, $position)
 	{
+		$accepts_card_type = CardTypes::USER_STORY;
+		$is_goalable = true;
+
+		if($key === BoardListsKeys::NOT_PRIORITIZED) {
+			$accepts_card_type = CardTypes::NOT_PRIORITIZED;
+			$is_goalable = false;
+		}
+
 		BoardList::create([
 			'name' => $this->getBoardListLabel($key, $workspace),
 			'key' => $this->getBoardListKey($key, $workspace),
-			'accepts_card_type' => 'user-story',
+			'accepts_card_type' => $accepts_card_type,
+			'is_goalable' => $is_goalable,
 			'position' => $position,
 		]);
 	}
@@ -101,7 +111,10 @@ class WorkspaceController extends Controller
 	{
 		$board_list = BoardList::where('key', $this->getBoardListKey($key, $workspace))->first();
 
-		$board_list->delete();
+		if(!empty($board_list)) {
+			$board_list->delete();
+		}
+
 	}
 
 	private function getBoardListLabel($key, $workspace)
